// src/enums/entity-name.enum.ts
export enum EntityNameE {
  archer = "Archer",
  armorer = "Armorer",
  fighter = "Fighter",
  looter = "Looter",
  loyalCompanion = "Loyal Companion",
  mule = "Mule",
  petSitter = "Pet Sitter",
  rat = "Rat",
  scholar = "Scholar",
  squire = "Squire",
  poisonousRat = "Poisonous Rat",
  bard = "Bard",
  berserker = "Berserker",
  cat = "Cat",
  druid = "Druid",
  innkeeper = "Innkeeper",
  loneWolf = "Lone Wolf",
  necromancer = "Necromancer",
  spiderEgg = "Spider Egg",
  summoner = "Summoner",
  trader = "Trader",
  spiderling = "Spiderling",
  risenSkeleton = "Risen Skeleton",
  alchemist = "Alchemist",
  blueJay = "Blue Jay",
  commander = "Commander",
  conjurer = "Conjurer",
  dragonEgg = "Dragon Egg",
  mentor = "Mentor",
  panda = "Panda",
  runeSmith = "Rune Smith",
  salvager = "Salvager",
  shaman = "Shaman",
  redImp = "Red Imp",
  greenImp = "Green Imp",
  blueImp = "Blue Imp",
  yellowImp = "Yellow Imp",
  dragon = "Dragon"
}